Green hr

+++ server/controllers/admin/userController.js
@@ -1,4 +1,11 @@
+const blog = require("../../models/blog");
+const question = require("../../models/question");
+const comment = require("../../models/comment");
+
 const User = require("./../../models/user");
+const blogcomment = require("../../models/blogcomment");
+const bloglike = require("../../models/bloglike");
+const like=require("../../models/like");
 
 
 // ===================================================
@@ -31,8 +38,25 @@ exports.changeRole = async (req,res)=>{
 
 exports.deleteUser = async (req,res)=>{
 	try{
+
+		var userdata= await User.find({_id:req.params.id});
+		if(userdata.role==="user")
+		{
+
 		var data = await User.findOneAndRemove({_id:req.params.id});
-		
+		 await question.deleteMany({user_id:req.params.id});
+		 await blog.deleteMany({user_id:req.params.id});
+		 await comment.deleteMany({user_id:req.params.id});
+		 await blogcomment.deleteMany({user_id:req.params.id});
+		 await bloglike.deleteMany({user_id:req.params.id});
+		 await like.deleteMany({user_id:req.params.id});
+		}
+		else
+		{
+			
+			return res.json({status:"OK",message:"you can't delete this superuser or admin"});
+		}
+
 		if(!data){
 			return res.json({status:"NOT_EXIST",message:"User does not exist."})
 		}
